move version control header to page header
