+++ src/studio/src/designer/frontend/app-development/layout/AppBar/AppBar.test.tsx
