+++ src/studio/src/designer/frontend/app-development/layout/AppBar/appBarConfig.tsx
