+++ src/studio/src/designer/frontend/app-development/layout/AppBar/index.ts
@@ -1,0 +1,1 @@
+export { AppBar } from './AppBar';
